+++ apps/src/lib/node/ledger/mod.rs
@@ -1,425 +1,199 @@
 pub mod protocol;
+mod shell;
 pub mod storage;
-mod tendermint;
+mod tendermint_node;
 
-use std::convert::TryFrom;
-use std::path::Path;
-use std::sync::mpsc;
+use std::convert::{TryFrom, TryInto};
+use std::future::Future;
+use std::pin::Pin;
+use std::sync::mpsc::channel;
+use std::task::{Context, Poll};
 
-use anoma_shared::ledger::gas::{self, BlockGasMeter};
-use anoma_shared::ledger::native_vp;
-use anoma_shared::ledger::storage::write_log::WriteLog;
-use anoma_shared::ledger::storage::MerkleRoot;
-use anoma_shared::proto::{self, Tx};
-use anoma_shared::types::address::Address;
-use anoma_shared::types::key::ed25519::PublicKey;
-use anoma_shared::types::storage::{BlockHash, BlockHeight, Key};
-use anoma_shared::types::token::Amount;
-use anoma_shared::types::{address, key, token};
-use borsh::BorshSerialize;
-use itertools::Itertools;
-use thiserror::Error;
+use anoma_shared::types::storage::{BlockHash, BlockHeight};
+use futures::future::FutureExt;
+use signal_hook::consts::TERM_SIGNALS;
+use signal_hook::iterator::Signals;
+use tendermint_proto::abci::CheckTxType;
+use tower::{Service, ServiceBuilder};
+use tower_abci::{response, split, BoxError, Request, Response, Server};
 
-use self::tendermint::{AbciMsg, AbciReceiver};
-use crate::{config, wallet};
+use crate::node::ledger::shell::{MempoolTxType, Shell};
+use crate::{config, genesis};
 
-#[derive(Error, Debug)]
-pub enum Error {
-    #[error("Error removing the DB data: {0}")]
-    RemoveDB(std::io::Error),
-    #[error("chain ID mismatch: {0}")]
-    ChainIdError(String),
-    #[error("Shell ABCI channel receiver error: {0}")]
-    AbciChannelRecvError(mpsc::RecvError),
-    #[error("Shell ABCI channel sender error: {0}")]
-    AbciChannelSendError(String),
-    #[error("Error decoding a transaction from bytes: {0}")]
-    TxDecodingError(proto::Error),
-    #[error("Error trying to apply a transaction: {0}")]
-    TxError(protocol::Error),
+impl Service<Request> for Shell {
+    type Error = BoxError;
+    type Future = Pin<
+        Box<dyn Future<Output = Result<Response, BoxError>> + Send + 'static>,
+    >;
+    type Response = Response;
+
+    fn poll_ready(
+        &mut self,
+        _cx: &mut Context<'_>,
+    ) -> Poll<Result<(), Self::Error>> {
+        Poll::Ready(Ok(()))
+    }
+
+    fn call(&mut self, req: Request) -> Self::Future {
+        tracing::info!(?req);
+        let rsp = match req {
+            Request::InitChain(init) => {
+                match self.init_chain(init) {
+                    Ok(mut resp) => {
+                        // Set the initial validator set
+                        let genesis = genesis::genesis();
+                        let mut abci_validator =
+                            tendermint_proto::abci::ValidatorUpdate::default();
+                        let pub_key = tendermint_proto::crypto::PublicKey {
+                            sum: Some(tendermint_proto::crypto::public_key::Sum::Ed25519(
+                                genesis.validator.keypair.public.to_bytes().to_vec(),
+                            )),
+                        };
+                        abci_validator.pub_key = Some(pub_key);
+                        abci_validator.power = genesis
+                            .validator
+                            .voting_power
+                            .try_into()
+                            .expect("unexpected validator's voting power");
+                        resp.validators.push(abci_validator);
+                        Ok(Response::InitChain(resp))
+                    }
+                    Err(inner) => Err(inner),
+                }
+            }
+            Request::Info(_) => Ok(Response::Info(self.last_state())),
+            Request::Query(query) => Ok(Response::Query(self.query(query))),
+            Request::BeginBlock(block) => {
+                match (
+                    BlockHash::try_from(&*block.hash),
+                    BlockHeight::try_from(
+                        block.header.expect("missing block's header").height,
+                    ),
+                ) {
+                    (Ok(hash), Ok(height)) => {
+                        let _ = self.begin_block(hash, height);
+                    }
+                    (Ok(_), Err(msg)) => {
+                        tracing::error!("Unexpected block height {}", msg);
+                    }
+                    (err @ Err(_), _) => tracing::error!("{:#?}", err),
+                };
+                Ok(Response::BeginBlock(Default::default()))
+            }
+            Request::DeliverTx(deliver_tx) => {
+                Ok(Response::DeliverTx(self.apply_tx(deliver_tx)))
+            }
+            Request::EndBlock(end) => match BlockHeight::try_from(end.height) {
+                Ok(height) => Ok(Response::EndBlock(self.end_block(height))),
+                Err(_) => {
+                    tracing::error!("Unexpected block height {}", end.height);
+                    Ok(Response::EndBlock(Default::default()))
+                }
+            },
+            Request::Commit(_) => Ok(Response::Commit(self.commit())),
+            Request::Flush(_) => Ok(Response::Flush(Default::default())),
+            Request::SetOption(_) => {
+                Ok(Response::SetOption(Default::default()))
+            }
+            Request::Echo(msg) => Ok(Response::Echo(response::Echo {
+                message: msg.message,
+            })),
+            Request::CheckTx(tx) => {
+                let r#type = match CheckTxType::from_i32(tx.r#type)
+                    .expect("received unexpected CheckTxType from ABCI")
+                {
+                    CheckTxType::New => MempoolTxType::NewTransaction,
+                    CheckTxType::Recheck => MempoolTxType::RecheckTransaction,
+                };
+                Ok(Response::CheckTx(self.mempool_validate(&*tx.tx, r#type)))
+            }
+            Request::ListSnapshots(_) => {
+                Ok(Response::ListSnapshots(Default::default()))
+            }
+            Request::OfferSnapshot(_) => {
+                Ok(Response::OfferSnapshot(Default::default()))
+            }
+            Request::LoadSnapshotChunk(_) => {
+                Ok(Response::LoadSnapshotChunk(Default::default()))
+            }
+            Request::ApplySnapshotChunk(_) => {
+                Ok(Response::ApplySnapshotChunk(Default::default()))
+            }
+        };
+        tracing::info!(?rsp);
+        Box::pin(async move { rsp.map_err(|e| e.into()) }.boxed())
+    }
 }
 
-pub type Result<T> = std::result::Result<T, Error>;
+pub fn reset(config: config::Ledger) -> Result<(), shell::Error> {
+    shell::reset(config)
+}
 
-pub fn run(config: config::Ledger) -> Result<()> {
-    // open a channel between ABCI (the sender) and the shell (the receiver)
-    let (sender, receiver) = mpsc::channel();
-    let shell =
-        Shell::new(receiver, &config.db, config::DEFAULT_CHAIN_ID.to_owned());
+#[tokio::main]
+async fn run_shell(config: config::Ledger) {
+    // Construct our ABCI application.
+    let service = Shell::new(&config.db, config::DEFAULT_CHAIN_ID.to_owned());
+
+    // Split it into components.
+    let (consensus, mempool, snapshot, info) = split::service(service, 1);
+
+    // Hand those components to the ABCI server, but customize request behavior
+    // for each category
+    let server = Server::builder()
+        .consensus(consensus)
+        .snapshot(snapshot)
+        .mempool(
+            ServiceBuilder::new()
+                .load_shed()
+                .buffer(10)
+                .service(mempool),
+        )
+        .info(
+            ServiceBuilder::new()
+                .load_shed()
+                .buffer(100)
+                .rate_limit(50, std::time::Duration::from_secs(1))
+                .service(info),
+        )
+        .finish()
+        .unwrap();
+
+    // Run the server with the shell
+    server.listen(config.address).await.unwrap();
+}
+
+pub fn run(config: config::Ledger) {
     // Run Tendermint ABCI server in another thread
-    let _tendermint_handle = std::thread::spawn(move || {
-        if let Err(err) = tendermint::run(sender.clone(), config) {
+    let home_dir = config.tendermint.clone();
+    let socket_address = config.address.to_string();
+    // used for shutting down Tendermint node
+    let (sender, receiver) = channel();
+    // start Tendermint node
+    let _ = std::thread::spawn(move || {
+        if let Err(err) =
+            tendermint_node::run(home_dir, &socket_address, receiver)
+        {
             tracing::error!(
                 "Failed to start-up a Tendermint node with {}",
                 err
             );
-            sender.send(AbciMsg::Terminate).unwrap();
         }
     });
+    // start the shell + ABCI server
+    let _ = std::thread::spawn(move || {
+        run_shell(config);
+    });
     tracing::info!("Anoma ledger node started.");
-    shell.run()
-}
 
-pub fn reset(config: config::Ledger) -> Result<()> {
-    // simply nuke the DB files
-    let db_path = &config.db;
-    match std::fs::remove_dir_all(&db_path) {
-        Err(e) if e.kind() == std::io::ErrorKind::NotFound => (),
-        res => res.map_err(Error::RemoveDB)?,
-    };
-    // reset Tendermint state
-    tendermint::reset(config);
-    Ok(())
-}
-
-#[derive(Debug)]
-pub struct Shell {
-    abci: AbciReceiver,
-    storage: storage::PersistentStorage,
-    gas_meter: BlockGasMeter,
-    write_log: WriteLog,
-}
-
-#[derive(Clone, Debug)]
-pub enum MempoolTxType {
-    /// A transaction that has not been validated by this node before
-    NewTransaction,
-    /// A transaction that has been validated at some previous level that may
-    /// need to be validated again
-    RecheckTransaction,
-}
-
-impl Shell {
-    pub fn new(
-        abci: AbciReceiver,
-        db_path: impl AsRef<Path>,
-        chain_id: String,
-    ) -> Self {
-        let mut storage = storage::open(db_path, chain_id);
-        storage
-            .load_last_state()
-            .map_err(|e| {
-                tracing::error!("Cannot load the last state from the DB {}", e);
-            })
-            .expect("PersistentStorage cannot be initialized");
-
-        Self {
-            abci,
-            storage,
-            gas_meter: BlockGasMeter::default(),
-            write_log: WriteLog::default(),
+    // If a termination signal is received, shut down the tendermint node
+    let mut signals =
+        Signals::new(TERM_SIGNALS).expect("Failed to creat OS signal handlers");
+    for sig in signals.forever() {
+        if TERM_SIGNALS.contains(&sig) {
+            sender.send(true).unwrap();
+            break;
         }
     }
 
-    /// Run the shell in the current thread (blocking). This is the intended way
-    /// to interact with the shell. It will forward received commands to the
-    /// appropriate internal methods which we do not expose outward.
-    ///
-    /// N.B. This is intended to be called by third party software whose
-    /// correctness we assume, e.g. the Tendermint ABCI. We thus do not
-    /// duplicate checks on the validity of state transitions here.
-    pub fn run(mut self) -> Result<()> {
-        loop {
-            let msg = self.abci.recv().map_err(Error::AbciChannelRecvError)?;
-            match msg {
-                AbciMsg::GetInfo { reply } => {
-                    let result = self.last_state();
-                    reply.send(result).map_err(|e| {
-                        Error::AbciChannelSendError(format!("GetInfo {}", e))
-                    })?
-                }
-                AbciMsg::InitChain { reply, chain_id } => {
-                    self.init_chain(chain_id)?;
-                    reply.send(()).map_err(|e| {
-                        Error::AbciChannelSendError(format!("InitChain {}", e))
-                    })?
-                }
-                AbciMsg::MempoolValidate { reply, tx, r#type } => {
-                    let result = self
-                        .mempool_validate(&tx, r#type)
-                        .map_err(|e| format!("{}", e));
-                    reply.send(result).map_err(|e| {
-                        Error::AbciChannelSendError(format!(
-                            "MempoolValidate {}",
-                            e
-                        ))
-                    })?
-                }
-                AbciMsg::BeginBlock {
-                    reply,
-                    hash,
-                    height,
-                } => {
-                    self.begin_block(hash, height);
-                    reply.send(()).map_err(|e| {
-                        Error::AbciChannelSendError(format!("BeginBlock {}", e))
-                    })?
-                }
-                AbciMsg::ApplyTx { reply, tx } => {
-                    let (gas, result) = self.apply_tx(&tx);
-                    let result = result.map_err(|e| e.to_string());
-                    reply.send((gas, result)).map_err(|e| {
-                        Error::AbciChannelSendError(format!("ApplyTx {}", e))
-                    })?
-                }
-                AbciMsg::EndBlock { reply, height } => {
-                    self.end_block(height);
-                    reply.send(()).map_err(|e| {
-                        Error::AbciChannelSendError(format!("EndBlock {}", e))
-                    })?
-                }
-                AbciMsg::CommitBlock { reply } => {
-                    let result = self.commit();
-                    reply.send(result).map_err(|e| {
-                        Error::AbciChannelSendError(format!(
-                            "CommitBlock {}",
-                            e
-                        ))
-                    })?
-                }
-                AbciMsg::AbciQuery {
-                    reply,
-                    path,
-                    data,
-                    height: _,
-                    prove: _,
-                } => {
-                    if path == "dry_run_tx" {
-                        let result = self
-                            .dry_run_tx(&data)
-                            .map_err(|e| format!("{}", e));
-
-                        reply.send(result).map_err(|e| {
-                            Error::AbciChannelSendError(format!(
-                                "ApplyTx {}",
-                                e
-                            ))
-                        })?
-                    }
-                }
-                AbciMsg::Terminate => {
-                    tracing::info!("Shutting down Anoma node");
-                    break;
-                }
-            }
-        }
-        Ok(())
-    }
-
-    /// Create a new genesis for the chain with specified id. This includes
-    /// 1. A set of initial users and tokens
-    /// 2. Setting up the validity predicates for both users and tokens
-    /// 3. A matchmaker
-    fn init_chain(&mut self, chain_id: String) -> Result<()> {
-        let (current_chain_id, _) = self.storage.get_chain_id();
-        if current_chain_id != chain_id {
-            return Err(Error::ChainIdError(format!(
-                "Current chain ID: {}, Tendermint chain ID: {}",
-                current_chain_id, chain_id
-            )));
-        }
-
-        // Initialize because there is no block
-        let token_vp =
-            std::fs::read("wasm/vp_token.wasm").expect("cannot load token VP");
-        let user_vp =
-            std::fs::read("wasm/vp_user.wasm").expect("cannot load user VP");
-
-        // TODO load initial accounts from genesis
-
-        // temporary account addresses for testing, generated by the
-        // address.rs module
-        let alberto = Address::decode("a1qq5qqqqqg4znssfsgcurjsfhgfpy2vjyxy6yg3z98pp5zvp5xgersvfjxvcnx3f4xycrzdfkak0xhx")
-            .expect("The genesis address shouldn't fail decoding");
-        let bertha = Address::decode("a1qq5qqqqqxv6yydz9xc6ry33589q5x33eggcnjs2xx9znydj9xuens3phxppnwvzpg4rrqdpswve4n9")
-            .expect("The genesis address shouldn't fail decoding");
-        let christel = Address::decode("a1qq5qqqqqxsuygd2x8pq5yw2ygdryxs6xgsmrsdzx8pryxv34gfrrssfjgccyg3zpxezrqd2y2s3g5s")
-            .expect("The genesis address shouldn't fail decoding");
-        let users = vec![alberto, bertha, christel];
-
-        let tokens = vec![
-            address::xan(),
-            address::btc(),
-            address::eth(),
-            address::dot(),
-            address::schnitzel(),
-            address::apfel(),
-            address::kartoffel(),
-        ];
-
-        for token in &tokens {
-            // default tokens VPs for testing
-            let key = Key::validity_predicate(&token);
-            self.storage
-                .write(&key, token_vp.to_vec())
-                .expect("Unable to write token VP");
-        }
-
-        for (user, token) in users.iter().cartesian_product(tokens.iter()) {
-            // default user VPs for testing
-            self.storage
-                .write(&Key::validity_predicate(user), user_vp.to_vec())
-                .expect("Unable to write user VP");
-
-            // default user's tokens for testing
-            self.storage
-                .write(
-                    &token::balance_key(token, user),
-                    Amount::whole(1_000_000)
-                        .try_to_vec()
-                        .expect("encode token amount"),
-                )
-                .expect("Unable to set genesis balance");
-
-            // default user's public keys for testing
-            let pk_key = key::ed25519::pk_key(user);
-            let pk = PublicKey::from(wallet::key_of(user.encode()).public);
-            self.storage
-                .write(&pk_key, pk.try_to_vec().expect("encode public key"))
-                .expect("Unable to set genesis user public key");
-        }
-
-        // Temporary for testing, we have a fixed matchmaker account.  This
-        // account has a public key for signing matchmaker txs and verifying
-        // their signatures in its VP. The VP is the same as the user's VP,
-        // which simply checks the signature. We could consider using the
-        // same key as the intent gossip's p2p key.
-        let matchmaker = address::matchmaker();
-        let matchmaker_pk = key::ed25519::pk_key(&matchmaker);
-        self.storage
-            .write(
-                &matchmaker_pk,
-                wallet::matchmaker_pk()
-                    .try_to_vec()
-                    .expect("encode public key"),
-            )
-            .expect("Unable to set genesis user public key");
-        self.storage
-            .write(&Key::validity_predicate(&matchmaker), user_vp.to_vec())
-            .expect("Unable to write matchmaker VP");
-
-        // TODO pass in the genesis object
-        native_vp::init_genesis_storage(&mut self.storage);
-        Ok(())
-    }
-
-    /// Validate a transaction request. On success, the transaction will
-    /// included in the mempool and propagated to peers, otherwise it will be
-    /// rejected.
-    fn mempool_validate(
-        &self,
-        tx_bytes: &[u8],
-        r#_type: MempoolTxType,
-    ) -> Result<()> {
-        let _tx = Tx::try_from(tx_bytes).map_err(Error::TxDecodingError)?;
-        Ok(())
-    }
-
-    /// Validate and apply a transaction.
-    fn apply_tx(
-        &mut self,
-        tx_bytes: &[u8],
-    ) -> (i64, Result<protocol::TxResult>) {
-        let result = protocol::apply_tx(
-            tx_bytes,
-            &mut self.gas_meter,
-            &mut self.write_log,
-            &self.storage,
-        )
-        .map_err(Error::TxError);
-
-        match result {
-            Ok(result) => {
-                if result.is_accepted() {
-                    tracing::info!(
-                        "all VPs accepted apply_tx storage modification {:#?}",
-                        result
-                    );
-                    self.write_log.commit_tx();
-                } else {
-                    tracing::info!(
-                        "some VPs rejected apply_tx storage modification {:#?}",
-                        result.vps_result.rejected_vps
-                    );
-                    self.write_log.drop_tx();
-                }
-
-                let gas = gas::as_i64(result.gas_used);
-                (gas, Ok(result))
-            }
-            err @ Err(_) => {
-                let gas =
-                    gas::as_i64(self.gas_meter.get_current_transaction_gas());
-                (gas, err)
-            }
-        }
-    }
-
-    /// Simulate validation and application of a transaction.
-    fn dry_run_tx(&mut self, tx_bytes: &[u8]) -> Result<String> {
-        let mut gas_meter = BlockGasMeter::default();
-        let mut write_log = self.write_log.clone();
-        let result = protocol::apply_tx(
-            tx_bytes,
-            &mut gas_meter,
-            &mut write_log,
-            &self.storage,
-        )
-        .map_err(Error::TxError)?;
-        Ok(result.to_string())
-    }
-
-    /// Begin a new block.
-    fn begin_block(&mut self, hash: BlockHash, height: BlockHeight) {
-        self.gas_meter.reset();
-        self.storage.begin_block(hash, height).unwrap();
-    }
-
-    /// End a block.
-    fn end_block(&mut self, _height: BlockHeight) {}
-
-    /// Commit a block. Persist the application state and return the Merkle root
-    /// hash.
-    fn commit(&mut self) -> MerkleRoot {
-        // commit changes from the write-log to storage
-        self.write_log
-            .commit_block(&mut self.storage)
-            .expect("Expected committing block write log success");
-        // store the block's data in DB
-        // TODO commit async?
-        self.storage.commit().unwrap_or_else(|e| {
-            tracing::error!(
-                "Encountered a storage error while committing a block {:?}",
-                e
-            )
-        });
-        let root = self.storage.merkle_root();
-        tracing::info!(
-            "Committed block hash: {}, height: {}",
-            root,
-            self.storage.current_height,
-        );
-        root
-    }
-
-    /// Load the Merkle root hash and the height of the last committed block, if
-    /// any.
-    fn last_state(&self) -> Option<(MerkleRoot, u64)> {
-        let result = self.storage.get_state();
-        match &result {
-            Some((root, height)) => {
-                tracing::info!(
-                    "Last state root hash: {}, height: {}",
-                    root,
-                    height
-                )
-            }
-            None => {
-                tracing::info!(
-                    "No state could be found, chain is not initialized"
-                )
-            }
-        }
-        result
-    }
+    tracing::info!("Shutting down Anoma node");
 }
